Instrument location and scaling now uses the dimensions of the background image after resizing to fit the available window size. Allows for more accrate positioning and retains image aspect ratio

+++ RemoteCockpitClasses/InstrumentLocation.cs
@@ -8,9 +8,9 @@ namespace RemoteCockpitClasses
     public class InstrumentLocation
     {
         public InstrumentType Type { get; set; }
-        public int Top { get; set; }
-        public int Left { get; set; }
-        public int Height { get; set; }
-        public int Width { get; set; }
+        public double Top { get; set; }
+        public double Left { get; set; }
+        public double Height { get; set; }
+        public double Width { get; set; }
     }
 }
